requirements.txt
Django>=5.0
django-colorfield>=0.11.0 